# artifacts/api-server/src/routes/analytics.ts
import { Router } from "express";
import { db } from "@workspace/db";
import { bookings, workspaces, activityEvents } from "@workspace/db";
import { eq, desc, gte, and } from "drizzle-orm";
import {
  GetOccupancyAnalyticsQueryParams,
  GetRevenueAnalyticsQueryParams,
  GetActivityFeedQueryParams,
} from "@workspace/api-zod";

const router = Router();

const NGN_PER_USDC = 1600;

router.get("/dashboard", async (req, res) => {
  try {
    const allBookings = await db.select().from(bookings);
    const allWorkspaces = await db.select().from(workspaces);

    const activeBookings = allBookings.filter((b) => b.status === "active");
    const completedBookings = allBookings.filter((b) => b.status === "completed");

    const today = new Date();
    today.setHours(0, 0, 0, 0);

    const todayBookings = allBookings.filter(
      (b) => b.checkInTime && b.checkInTime >= today
    );

    const totalRevenueUsdc = completedBookings.reduce(
      (sum, b) => sum + (b.billedAmountUsdc ?? 0),
      0
    );

    const occupancyRate =
      allWorkspaces.length > 0
        ? activeBookings.length / allWorkspaces.length
        : 0;

    const avgDurationSecs =
      completedBookings.length > 0
        ? completedBookings.reduce(
            (sum, b) => sum + (b.actualDurationSeconds ?? 0),
            0
          ) / completedBookings.length
        : 0;

    const yieldEarned = totalRevenueUsdc * 0.05 * (1 / 365);

    const workspaceSessionCounts: Record<string, number> = {};
    allBookings.forEach((b) => {
      workspaceSessionCounts[b.workspaceId] =
        (workspaceSessionCounts[b.workspaceId] ?? 0) + 1;
    });

    const topWorkspaceId = Object.entries(workspaceSessionCounts).sort(
      ([, a], [, b]) => b - a
    )[0]?.[0];

    const topWorkspace = allWorkspaces.find((w) => w.id === topWorkspaceId);

    const workspaceStats = allWorkspaces.map((ws) => {
      const wsSessions = todayBookings.filter((b) => b.workspaceId === ws.id);
      const wsRevenue = wsSessions.reduce(
        (sum, b) => sum + (b.billedAmountUsdc ?? 0),
        0
      );
      const wsActive = activeBookings.filter(
        (b) => b.workspaceId === ws.id
      ).length;

      return {
        workspace_id: ws.id,
        workspace_name: ws.name,
        sessions_today: wsSessions.length,
        revenue_today_usdc: wsRevenue,
        occupancy_rate: wsActive > 0 ? 1 : 0,
      };
    });

    res.json({
      total_revenue_usdc: parseFloat(totalRevenueUsdc.toFixed(4)),
      total_revenue_ngn: parseFloat((totalRevenueUsdc * NGN_PER_USDC).toFixed(2)),
      total_sessions_today: todayBookings.length,
      active_sessions: activeBookings.length,
      occupancy_rate: parseFloat(occupancyRate.toFixed(4)),
      avg_session_duration_hours: parseFloat((avgDurationSecs / 3600).toFixed(2)),
      top_workspace: topWorkspace?.name ?? "N/A",
      yield_earned_usdc: parseFloat(yieldEarned.toFixed(6)),
      revenue_trend: completedBookings.length > 2 ? "up" : "flat",
      revenue_change_pct: 12.5,
      workspace_stats: workspaceStats,
    });
  } catch (err) {
    console.error(err);
    res.status(500).json({ error: "Failed to get dashboard" });
  }
});

router.get("/occupancy", async (req, res) => {
  try {
    const query = GetOccupancyAnalyticsQueryParams.safeParse(req.query);
    const days = query.success ? (query.data.days ?? 30) : 30;

    const allWorkspaces = await db.select().from(workspaces);
    const allBookings = await db.select().from(bookings);

    const result = [];
    for (let d = days - 1; d >= 0; d--) {
      const date = new Date();
      date.setDate(date.getDate() - d);
      date.setHours(0, 0, 0, 0);
      const dateStr = date.toISOString().split("T")[0];

      for (const ws of allWorkspaces) {
        const dayBookings = allBookings.filter((b) => {
          if (!b.checkInTime) return false;
          const bDate = new Date(b.checkInTime);
          bDate.setHours(0, 0, 0, 0);
          return bDate.getTime() === date.getTime() && b.workspaceId === ws.id;
        });

        const totalHours = dayBookings.reduce(
          (sum, b) => sum + (b.plannedDurationHours ?? 0),
          0
        );

        result.push({
          date: dateStr,
          workspace_id: ws.id,
          workspace_name: ws.name,
          occupancy_rate: Math.min(1, totalHours / 15),
          total_hours: totalHours,
        });
      }
    }

    res.json(result);
  } catch (err) {
    res.status(500).json({ error: "Failed to get occupancy" });
  }
});

router.get("/revenue", async (req, res) => {
  try {
    const query = GetRevenueAnalyticsQueryParams.safeParse(req.query);
    const days = query.success ? (query.data.days ?? 30) : 30;

    const allWorkspaces = await db.select().from(workspaces);
    const allBookings = await db.select().from(bookings);

    const result = [];
    for (let d = days - 1; d >= 0; d--) {
      const date = new Date();
      date.setDate(date.getDate() - d);
      date.setHours(0, 0, 0, 0);
      const dateStr = date.toISOString().split("T")[0];

      for (const ws of allWorkspaces) {
        const dayBookings = allBookings.filter((b) => {
          if (!b.checkInTime) return false;
          const bDate = new Date(b.checkInTime);
          bDate.setHours(0, 0, 0, 0);
          return (
            bDate.getTime() === date.getTime() &&
            b.workspaceId === ws.id &&
            b.status === "completed"
          );
        });

        const revenueUsdc = dayBookings.reduce(
          (sum, b) => sum + (b.billedAmountUsdc ?? 0),
          0
        );

        result.push({
          date: dateStr,
          workspace_id: ws.id,
          workspace_name: ws.name,
          revenue_usdc: parseFloat(revenueUsdc.toFixed(4)),
          revenue_ngn: parseFloat((revenueUsdc * NGN_PER_USDC).toFixed(2)),
          session_count: dayBookings.length,
        });
      }
    }

    res.json(result);
  } catch (err) {
    res.status(500).json({ error: "Failed to get revenue" });
  }
});

router.get("/activity", async (req, res) => {
  try {
    const query = GetActivityFeedQueryParams.safeParse(req.query);
    const limit = query.success ? (query.data.limit ?? 20) : 20;

    const events = await db
      .select()
      .from(activityEvents)
      .orderBy(desc(activityEvents.createdAt))
      .limit(limit);

    res.json(
      events.map((e) => ({
        id: e.id,
        event_type: e.eventType,
        workspace_id: e.workspaceId,
        workspace_name: e.workspaceName,
        user_name: e.userName,
        amount_usdc: e.amountUsdc,
        amount_ngn: e.amountNgn,
        timestamp: e.createdAt.toISOString(),
        metadata: JSON.parse(e.metadata),
      }))
    );
  } catch (err) {
    res.status(500).json({ error: "Failed to get activity feed" });
  }
});

export default router;