# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JaIre Web3 Coworking Platform API
servers:
  - url: /api
    description: Baire AI concierge (Node api-server)
...
paths:
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List bookings for the current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, cancelled, pending]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: User bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking / check-in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a booking by ID
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Booking detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}/checkout:
    post:
      operationId: checkoutBooking
      tags: [bookings]
      summary: Check out of a booking (settle escrow)
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Checkout summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSummary"
  /bookings/{bookingId}/status:
    get:
      operationId: getBookingStatus
      tags: [bookings]
      summary: Real-time ticking status for an active booking
      parameters:
        - name: bookingId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Live session status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionStatus"
  /wallet/connect:
    post:
      operationId: connectWallet
      tags: [wallet]
      summary: Connect or create invisible Web3Auth MPC wallet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletConnectBody"
      responses:
        "200":
          description: Wallet info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletInfo"
...
  /wallet/fund:
    post:
      operationId: fundWallet
      tags: [wallet]
      summary: Fund wallet via fiat (NGN → USDC via Paystack)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FundWalletBody"
      responses:
        "200":
          description: Funding initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FundWalletResponse"
...
  /analytics/occupancy:
    get:
      operationId: getOccupancyAnalytics
      tags: [analytics]
      summary: Occupancy rates per workspace per day
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Occupancy data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OccupancyData"
  /analytics/revenue:
    get:
      operationId: getRevenueAnalytics
      tags: [analytics]
      summary: Revenue breakdown by workspace and time
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Revenue data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RevenueData"
  /analytics/activity:
    get:
      operationId: getActivityFeed
      tags: [analytics]
      summary: Recent activity feed across all workspaces
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Activity events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityEvent"
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: string
        workspace_id:
          type: string
        workspace_name:
          type: string
        user_id:
          type: string
        status:
          type: string
          enum: [pending, active, completed, cancelled]
        check_in_time:
          type: string
          format: date-time
        check_out_time:
          type: string
          format: date-time
        planned_duration_hours:
          type: number
        actual_duration_seconds:
          type: integer
        escrow_amount_usdc:
          type: number
        billed_amount_usdc:
          type: number
        refunded_amount_usdc:
          type: number
        escrow_account:
          type: string
          description: Solana escrow account pubkey
        transaction_signature:
          type: string
        payment_method:
          type: string
          enum: [paystack, roqqu, usdc_wallet]
        ngn_amount_paid:
          type: number
      required: [id, workspace_id, status]
    CreateBookingBody:
      type: object
      properties:
        workspace_id:
          type: string
        planned_duration_hours:
          type: number
        payment_method:
          type: string
          enum: [paystack, roqqu, usdc_wallet]
        web3auth_token:
          type: string
        user_email:
          type: string
        user_name:
          type: string
      required: [workspace_id, planned_duration_hours]
    CheckoutSummary:
      type: object
      properties:
        booking_id:
          type: string
        workspace_name:
          type: string
        duration_seconds:
          type: integer
        duration_display:
          type: string
        billed_usdc:
          type: number
        refunded_usdc:
          type: number
        yield_earned_usdc:
          type: number
        transaction_signature:
          type: string
        check_in_time:
          type: string
          format: date-time
        check_out_time:
          type: string
          format: date-time
      required: [booking_id, duration_seconds, billed_usdc]
    SessionStatus:
      type: object
      properties:
        booking_id:
          type: string
        workspace_id:
          type: string
        workspace_name:
          type: string
        status:
          type: string
          enum: [active, completed, cancelled]
        elapsed_seconds:
          type: integer
        planned_seconds:
          type: integer
        elapsed_display:
          type: string
        current_cost_usdc:
          type: number
        current_cost_ngn:
          type: number
        escrowed_usdc:
          type: number
        check_in_time:
          type: string
          format: date-time
      required: [booking_id, status, elapsed_seconds, current_cost_usdc]
    WalletConnectBody:
      type: object
      properties:
        web3auth_token:
          type: string
        email:
          type: string
      required: [web3auth_token]
    WalletInfo:
      type: object
      properties:
        public_key:
          type: string
        wallet_type:
          type: string
          enum: [mpc, custodial, external]
        balance_usdc:
          type: number
        balance_sol:
          type: number
        verifier_id:
          type: string
        connected:
          type: boolean
      required: [public_key, connected]
...
    FundWalletBody:
      type: object
      properties:
        amount_ngn:
          type: number
          description: Amount in Nigerian Naira to convert to USDC
        payment_method:
          type: string
          enum: [paystack, roqqu]
        email:
          type: string
      required: [amount_ngn, payment_method, email]
    FundWalletResponse:
      type: object
      properties:
        payment_url:
          type: string
          description: Paystack/Roqqu checkout URL
        reference:
          type: string
        amount_ngn:
          type: number
        estimated_usdc:
          type: number
        exchange_rate:
          type: number
          description: NGN per USDC
      required: [payment_url, reference, amount_ngn, estimated_usdc]
...
    OccupancyData:
      type: object
      properties:
        date:
          type: string
          format: date
        workspace_id:
          type: string
        workspace_name:
          type: string
        occupancy_rate:
          type: number
        total_hours:
          type: number
      required: [date, workspace_id, occupancy_rate]
    RevenueData:
      type: object
      properties:
        date:
          type: string
          format: date
        workspace_id:
          type: string
        workspace_name:
          type: string
        revenue_usdc:
          type: number
        revenue_ngn:
          type: number
        session_count:
          type: integer
      required: [date, workspace_id, revenue_usdc]
    ActivityEvent:
      type: object
      properties:
        id:
          type: string
        event_type:
          type: string
          enum: [check_in, check_out, payment, wallet_funded, booking_cancelled]
        workspace_id:
          type: string
        workspace_name:
          type: string
        user_name:
          type: string
        amount_usdc:
          type: number
        amount_ngn:
          type: number
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object
      required: [id, event_type, timestamp]