# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JaIre Web3 Coworking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingPaymentMethod } from "./bookingPaymentMethod";
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: string;
  workspace_id: string;
  workspace_name?: string;
  user_id?: string;
  status: BookingStatus;
  check_in_time?: Date;
  check_out_time?: Date;
  planned_duration_hours?: number;
  actual_duration_seconds?: number;
  escrow_amount_usdc?: number;
  billed_amount_usdc?: number;
  refunded_amount_usdc?: number;
  /** Solana escrow account pubkey */
  escrow_account?: string;
  transaction_signature?: string;
  payment_method?: BookingPaymentMethod;
  ngn_amount_paid?: number;
}