# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JaIre Web3 Coworking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookingBodyPaymentMethod } from "./createBookingBodyPaymentMethod";

export interface CreateBookingBody {
  workspace_id: string;
  planned_duration_hours: number;
  payment_method?: CreateBookingBodyPaymentMethod;
  web3auth_token?: string;
  user_email?: string;
  user_name?: string;
}