# lib/api-zod/src/generated/types/fundWalletBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JaIre Web3 Coworking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { FundWalletBodyPaymentMethod } from "./fundWalletBodyPaymentMethod";

export interface FundWalletBody {
  /** Amount in Nigerian Naira to convert to USDC */
  amount_ngn: number;
  payment_method: FundWalletBodyPaymentMethod;
  email: string;
}