# lib/api-zod/src/generated/types/walletInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * JaIre Web3 Coworking Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { WalletInfoWalletType } from "./walletInfoWalletType";

export interface WalletInfo {
  public_key: string;
  wallet_type?: WalletInfoWalletType;
  balance_usdc?: number;
  balance_sol?: number;
  verifier_id?: string;
  connected: boolean;
}